Fix native buffer view port gates
