+++ crates/perry-codegen/tests/native_proof_regressions.rs
@@ -704,7 +704,7 @@ fn artifact_schema_v6_records_pod_dynamic_write_fallback() {
 }
 
 #[test]
-fn artifact_schema_v6_rejects_inexact_pod_initializer_values() {
+fn artifact_schema_v8_rejects_inexact_pod_initializer_values() {
     let packet_ty = pod_type(&[
         ("tag", Type::Named("PerryU32".to_string())),
         ("gain", Type::Named("PerryF32".to_string())),
@@ -728,7 +728,7 @@ fn artifact_schema_v6_rejects_inexact_pod_initializer_values() {
     ];
 
     let artifact = compile_artifact_json("artifact_c_layout_pod_init_reject.ts", body);
-    assert_eq!(artifact["schema_version"], 6);
+    assert_eq!(artifact["schema_version"], 8);
     assert_eq!(artifact["summary"]["pod_layout_count"], 0);
     assert_eq!(artifact["summary"]["pod_record_count"], 0);
     assert!(artifact["pod_layouts"].as_array().unwrap().is_empty());
@@ -798,766 +798,6 @@ fn artifact_schema_v6_records_pod_pointerful_field_rejection() {
                     })
             }),
         "expected explicit pointerful POD rejection record:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn artifact_records_buffer_read_u32_and_unsigned_materialization() {
-    let body = vec![
-        buffer_let(1, "buf", int(8)),
-        Stmt::Return(Some(Expr::Call {
-            callee: Box::new(Expr::PropertyGet {
-                object: Box::new(local(1)),
-                property: "readUInt32BE".to_string(),
-            }),
-            args: vec![int(0)],
-            type_args: Vec::new(),
-        })),
-    ];
-
-    let artifact = compile_artifact_json("artifact_buffer_read_u32.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    assert!(
-        records.iter().any(|record| {
-            record["expr_kind"] == "BufferNumericRead"
-                && record["consumer"] == "BufferNumericRead.native_u32"
-                && record["native_rep_name"] == "u32"
-                && record["llvm_ty"] == "i32"
-                && record["native_value_state"] == "region_local"
-        }),
-        "expected region-local u32 buffer numeric read record:\n{artifact:#}"
-    );
-    assert!(
-        records.iter().any(|record| {
-            record["consumer"] == "materialize_js_value"
-                && record["native_value_state"] == "materialized"
-                && record["native_abi_transition"]["from_native_rep"] == "u32"
-                && record["native_abi_transition"]["to_native_rep"] == "js_value"
-                && record["native_abi_transition"]["op"] == "unsigned_int_to_float"
-                && record["native_abi_transition"]["lossy"] == false
-        }),
-        "expected unsigned u32 JS materialization record:\n{artifact:#}"
-    );
-    assert!(
-        artifact["summary"]["native_abi_transition_count"]
-            .as_u64()
-            .is_some_and(|count| count >= 1)
-            && artifact["summary"]["native_abi_transition_op_counts"]["unsigned_int_to_float"]
-                .as_u64()
-                .is_some_and(|count| count >= 1),
-        "expected transition summary counts for unsigned materialization:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn loop_length_bound_does_not_prove_multibyte_buffer_read_inbounds() {
-    let body = vec![
-        buffer_let(1, "buf", int(8)),
-        for_loop(
-            2,
-            length(1),
-            vec![Stmt::Expr(call(
-                Expr::PropertyGet {
-                    object: Box::new(local(1)),
-                    property: "readUInt32BE".to_string(),
-                },
-                vec![local(2)],
-            ))],
-        ),
-        Stmt::Return(Some(int(0))),
-    ];
-
-    let ir = compile_ir("loop_bound_multibyte_buffer_read.ts", body.clone());
-    assert!(
-        !ir.contains("getelementptr inbounds i8"),
-        "`i < buf.length` only proves one-byte Buffer access; multi-byte reads must not emit an inbounds GEP:\n{ir}"
-    );
-
-    let artifact = compile_artifact_json("artifact_loop_bound_multibyte_buffer_read.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    assert!(
-        !records.iter().any(|record| {
-            record["expr_kind"] == "BufferNumericRead"
-                && record["consumer"] == "BufferNumericRead.native_u32"
-        }),
-        "multi-byte Buffer read must not consume a one-byte loop proof:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn artifact_records_buffer_read_double_as_f64() {
-    let body = vec![
-        buffer_let(1, "buf", int(8)),
-        Stmt::Return(Some(Expr::Call {
-            callee: Box::new(Expr::PropertyGet {
-                object: Box::new(local(1)),
-                property: "readDoubleLE".to_string(),
-            }),
-            args: vec![int(0)],
-            type_args: Vec::new(),
-        })),
-    ];
-
-    let artifact = compile_artifact_json("artifact_buffer_read_double.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    assert!(
-        records.iter().any(|record| {
-            record["expr_kind"] == "BufferNumericRead"
-                && record["consumer"] == "BufferNumericRead.native_f64"
-                && record["native_rep_name"] == "f64"
-                && record["llvm_ty"] == "double"
-                && record["native_value_state"] == "region_local"
-        }),
-        "expected region-local f64 buffer numeric read record:\n{artifact:#}"
-    );
-    assert!(
-        records.iter().any(|record| {
-            record["consumer"] == "materialize_js_value"
-                && record["native_abi_transition"]["from_native_rep"] == "f64"
-                && record["native_abi_transition"]["op"] == "none"
-                && record["native_abi_transition"]["lossy"] == false
-        }),
-        "expected no-op f64 JS materialization record:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn artifact_records_buffer_read_float_as_f32_and_float_extend_materialization() {
-    let body = vec![
-        buffer_let(1, "buf", int(8)),
-        Stmt::Return(Some(call(
-            Expr::PropertyGet {
-                object: Box::new(local(1)),
-                property: "readFloatLE".to_string(),
-            },
-            vec![int(0)],
-        ))),
-    ];
-
-    let artifact = compile_artifact_json("artifact_buffer_read_f32.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    assert!(
-        records.iter().any(|record| {
-            record["expr_kind"] == "BufferNumericRead"
-                && record["consumer"] == "BufferNumericRead.native_f32"
-                && record["native_rep_name"] == "f32"
-                && record["llvm_ty"] == "float"
-                && record["native_value_state"] == "region_local"
-        }),
-        "expected region-local f32 buffer numeric read record:\n{artifact:#}"
-    );
-    assert!(
-        records.iter().any(|record| {
-            record["consumer"] == "materialize_js_value"
-                && record["native_value_state"] == "materialized"
-                && record["native_abi_transition"]["from_native_rep"] == "f32"
-                && record["native_abi_transition"]["to_native_rep"] == "js_value"
-                && record["native_abi_transition"]["op"] == "float_extend"
-                && record["native_abi_transition"]["lossy"] == false
-        }),
-        "expected explicit f32->double materialization record:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn artifact_records_width_aware_buffer_numeric_read_facts() {
-    let reads = [
-        ("readUInt16BE", 2, "big", false, false, "u32"),
-        ("readUInt16LE", 2, "little", false, false, "u32"),
-        ("readUInt32BE", 4, "big", false, false, "u32"),
-        ("readUInt32LE", 4, "little", false, false, "u32"),
-        ("readInt16BE", 2, "big", true, false, "i32"),
-        ("readInt16LE", 2, "little", true, false, "i32"),
-        ("readInt32BE", 4, "big", true, false, "i32"),
-        ("readInt32LE", 4, "little", true, false, "i32"),
-        ("readFloatBE", 4, "big", false, true, "f32"),
-        ("readFloatLE", 4, "little", false, true, "f32"),
-        ("readDoubleBE", 8, "big", false, true, "f64"),
-        ("readDoubleLE", 8, "little", false, true, "f64"),
-    ];
-    let mut body = vec![buffer_let(1, "buf", int(16))];
-    for (method, ..) in reads {
-        body.push(Stmt::Expr(buffer_read(1, method, int(0))));
-    }
-    body.push(Stmt::Return(Some(int(0))));
-
-    let artifact = compile_artifact_json("artifact_width_aware_buffer_reads.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    for (method, width, endian, signed, floating, native_rep) in reads {
-        assert!(
-            records.iter().any(|record| {
-                record["expr_kind"] == "BufferNumericRead"
-                    && record["native_rep_name"] == native_rep
-                    && record["buffer_access"]["access_width_bytes"] == width
-                    && record["buffer_access"]["bounds_width_units"] == width
-                    && record["buffer_access"]["index_unit"] == "byte"
-                    && record["buffer_access"]["element_width_bytes"] == 1
-                    && record["buffer_access"]["endian"] == endian
-                    && record["buffer_access"]["signed"] == signed
-                    && record["buffer_access"]["floating"] == floating
-                    && record["notes"].as_array().is_some_and(|notes| {
-                        notes.iter().any(|note| {
-                            note.as_str()
-                                .is_some_and(|note| note == format!("method={method}"))
-                        })
-                    })
-            }),
-            "expected structured buffer_access facts for {method}:\n{artifact:#}"
-        );
-    }
-}
-
-#[test]
-fn loop_bound_by_buffer_length_only_does_not_prove_wide_read() {
-    let body = vec![
-        buffer_let(1, "buf", int(16)),
-        for_loop(
-            2,
-            length(1),
-            vec![Stmt::Expr(buffer_read(1, "readUInt32BE", local(2)))],
-        ),
-        Stmt::Return(Some(int(0))),
-    ];
-
-    let artifact = compile_artifact_json("artifact_wide_read_needs_width_guard.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    assert!(
-        !records.iter().any(|record| {
-            record["expr_kind"] == "BufferNumericRead"
-                && record["consumer"] == "BufferNumericRead.native_u32"
-                && record["buffer_access"]["access_width_bytes"] == 4
-        }),
-        "i < buf.length must not prove a 4-byte Buffer read:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn explicit_width_guard_proves_wide_buffer_read() {
-    let body = vec![
-        buffer_let(1, "buf", int(16)),
-        Stmt::For {
-            init: Some(Box::new(number_let(2, "i", true, int(0)))),
-            condition: Some(Expr::Compare {
-                op: CompareOp::Le,
-                left: Box::new(add(local(2), int(4))),
-                right: Box::new(length(1)),
-            }),
-            update: Some(increment(2)),
-            body: vec![Stmt::Expr(buffer_read(1, "readUInt32BE", local(2)))],
-        },
-        Stmt::Return(Some(int(0))),
-    ];
-
-    let artifact = compile_artifact_json("artifact_width_guard_buffer_read.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    assert!(
-        records.iter().any(|record| {
-            record["expr_kind"] == "BufferNumericRead"
-                && record["consumer"] == "BufferNumericRead.native_u32"
-                && record["bounds_state"]["guarded"]["guard_id"]
-                    .as_str()
-                    .is_some_and(|id| id.contains("width_4"))
-                && record["buffer_access"]["access_width_bytes"] == 4
-                && record["buffer_access"]["bounds_width_units"] == 4
-        }),
-        "expected i + 4 <= buf.length to guard a 4-byte native read:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn proven_buffer_and_typed_array_reads_are_numeric_operands() {
-    let body = vec![
-        buffer_let(1, "buf", int(16)),
-        number_let(2, "sum", true, int(0)),
-        Stmt::For {
-            init: Some(Box::new(number_let(3, "i", true, int(0)))),
-            condition: Some(Expr::Compare {
-                op: CompareOp::Le,
-                left: Box::new(add(local(3), int(4))),
-                right: Box::new(length(1)),
-            }),
-            update: Some(increment(3)),
-            body: vec![Stmt::Expr(Expr::LocalSet(
-                2,
-                Box::new(add(local(2), buffer_read(1, "readUInt32BE", local(3)))),
-            ))],
-        },
-        typed_array_let(
-            4,
-            "u16",
-            "Uint16Array",
-            perry_hir::TYPED_ARRAY_KIND_UINT16,
-            int(8),
-        ),
-        for_loop(
-            5,
-            int(8),
-            vec![Stmt::Expr(Expr::LocalSet(
-                2,
-                Box::new(add(local(2), index_get(4, local(5)))),
-            ))],
-        ),
-        Stmt::Return(Some(local(2))),
-    ];
-
-    let ir = compile_ir("numeric_buffer_typed_array_operands.ts", body);
-    assert!(
-        !ir.contains("call double @js_number_coerce"),
-        "proven native numeric reads should not force JS number coercion:\n{ir}"
-    );
-    assert!(
-        ir.contains("load i32, ptr") && ir.contains("align 1"),
-        "byte-offset Buffer numeric reads must use unaligned-safe loads:\n{ir}"
-    );
-}
-
-#[test]
-fn artifact_records_tracked_typed_array_native_reads() {
-    let arrays = [
-        (
-            1,
-            "u16",
-            "Uint16Array",
-            perry_hir::TYPED_ARRAY_KIND_UINT16,
-            "u32",
-            2,
-            false,
-            false,
-        ),
-        (
-            3,
-            "i32s",
-            "Int32Array",
-            perry_hir::TYPED_ARRAY_KIND_INT32,
-            "i32",
-            4,
-            true,
-            false,
-        ),
-        (
-            5,
-            "u32s",
-            "Uint32Array",
-            perry_hir::TYPED_ARRAY_KIND_UINT32,
-            "u32",
-            4,
-            false,
-            false,
-        ),
-        (
-            7,
-            "f32s",
-            "Float32Array",
-            perry_hir::TYPED_ARRAY_KIND_FLOAT32,
-            "f32",
-            4,
-            false,
-            true,
-        ),
-        (
-            9,
-            "f64s",
-            "Float64Array",
-            perry_hir::TYPED_ARRAY_KIND_FLOAT64,
-            "f64",
-            8,
-            false,
-            true,
-        ),
-    ];
-    let mut body = Vec::new();
-    for (array_id, name, class_name, kind, ..) in arrays {
-        body.push(typed_array_let(array_id, name, class_name, kind, int(8)));
-        body.push(for_loop(
-            array_id + 1,
-            int(8),
-            vec![Stmt::Expr(index_get(array_id, local(array_id + 1)))],
-        ));
-    }
-    body.push(Stmt::Return(Some(int(0))));
-
-    let artifact = compile_artifact_json("artifact_tracked_typed_array_reads.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    for (_, name, _, _, native_rep, width, signed, floating) in arrays {
-        assert!(
-            records.iter().any(|record| {
-                record["expr_kind"] == "TypedArrayGet"
-                    && record["native_rep_name"] == native_rep
-                    && record["access_mode"] == "unchecked_native"
-                    && record["buffer_access"]["index_unit"] == "element"
-                    && record["buffer_access"]["access_width_bytes"] == width
-                    && record["buffer_access"]["element_width_bytes"] == width
-                    && record["buffer_access"]["bounds_width_units"] == 1
-                    && record["buffer_access"]["signed"] == signed
-                    && record["buffer_access"]["floating"] == floating
-            }),
-            "expected native typed-array read record for {name}:\n{artifact:#}"
-        );
-    }
-    assert!(
-        records.iter().any(|record| {
-            record["consumer"] == "materialize_js_value"
-                && record["materialization_reason"] == "runtime_api"
-        }) && !records.iter().any(|record| {
-            record["consumer"] == "materialize_js_value"
-                && record["materialization_reason"] == "unknown_bounds"
-        }),
-        "proven typed-array loads should materialize at a JSValue boundary, not as an unknown-bounds hazard:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn artifact_records_native_owned_typed_array_facts() {
-    let body = vec![
-        native_arena_owner_let(1, "owner", int(64), false),
-        native_arena_view_let(
-            2,
-            "view",
-            1,
-            "Float64Array",
-            perry_hir::TYPED_ARRAY_KIND_FLOAT64,
-            int(0),
-            int(8),
-        ),
-        for_loop(
-            3,
-            int(8),
-            vec![array_set(2, local(3), add(local(3), number(0.5)))],
-        ),
-        for_loop(4, int(8), vec![Stmt::Expr(index_get(2, local(4)))]),
-        Stmt::Return(Some(int(0))),
-    ];
-
-    let ir = compile_ir("native_owned_typed_array_fast_path.ts", body.clone());
-    assert!(
-        !ir.contains("call double @js_typed_array_get")
-            && !ir.contains("call void @js_typed_array_set"),
-        "native-owned typed-array hot loops should avoid typed-array runtime get/set:\n{ir}"
-    );
-
-    let artifact = compile_artifact_json("artifact_native_owned_typed_array.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    let native_record = records.iter().find(|record| {
-        record["expr_kind"] == "TypedArrayGet"
-            && record["consumer"] == "TypedArrayGet.native_f64"
-            && record["access_mode"] == "unchecked_native"
-            && !record["native_owned_view"].is_null()
-    });
-    let Some(record) = native_record else {
-        panic!("expected native-owned f64 typed-array get record:\n{artifact:#}");
-    };
-    assert_eq!(record["native_owned_view"]["owner_local_id"], 1);
-    assert_eq!(record["native_owned_view"]["owner_root_state"], "rooted");
-    assert_eq!(record["native_owned_view"]["disposed_state"], "alive");
-    assert_eq!(record["native_owned_view"]["byte_offset"], 0);
-    assert_eq!(record["native_owned_view"]["byte_length"], 64);
-    assert_eq!(record["native_owned_view"]["element_width_bytes"], 8);
-    assert_eq!(record["native_owned_view"]["pointer_free_backing"], true);
-    assert_eq!(record["alias_state"], "no_alias_proven");
-    assert!(
-        records.iter().any(|record| {
-            record["expr_kind"] == "TypedArraySet"
-                && record["consumer"] == "f64_store"
-                && record["access_mode"] == "unchecked_native"
-                && record["alias_state"] == "no_alias_proven"
-                && !record["native_owned_view"].is_null()
-        }),
-        "expected native-owned f64 typed-array store record:\n{artifact:#}"
-    );
-    assert_eq!(artifact["summary"]["native_owned_view_count"], 4);
-}
-
-#[test]
-fn native_owned_typed_array_fallback_reasons_are_explicit() {
-    let disposed = compile_artifact_json(
-        "artifact_native_owned_disposed.ts",
-        vec![
-            native_arena_owner_let(1, "owner", int(64), false),
-            native_arena_view_let(
-                2,
-                "view",
-                1,
-                "Float64Array",
-                perry_hir::TYPED_ARRAY_KIND_FLOAT64,
-                int(0),
-                int(8),
-            ),
-            Stmt::Expr(Expr::NativeArenaDispose(Box::new(local(1)))),
-            Stmt::Return(Some(index_get(2, int(0)))),
-        ],
-    );
-    assert!(
-        disposed["records"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .any(|record| {
-                record["expr_kind"] == "TypedArrayGet"
-                    && record["consumer"] == "TypedArrayGet.slow_path"
-                    && record["access_mode"] == "dynamic_fallback"
-                    && record["materialization_reason"] == "use_after_dispose"
-                    && record["fallback_reason"] == "use_after_dispose"
-            }),
-        "expected disposed native-owned view fallback reason:\n{disposed:#}"
-    );
-
-    let stale_length = compile_artifact_json(
-        "artifact_native_owned_stale_length.ts",
-        vec![
-            native_arena_owner_let(1, "owner", int(64), false),
-            number_let(3, "len", true, int(8)),
-            native_arena_view_let(
-                2,
-                "view",
-                1,
-                "Float64Array",
-                perry_hir::TYPED_ARRAY_KIND_FLOAT64,
-                int(0),
-                local(3),
-            ),
-            Stmt::Expr(Expr::LocalSet(3, Box::new(int(4)))),
-            Stmt::Return(Some(index_get(2, int(0)))),
-        ],
-    );
-    assert!(
-        stale_length["records"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .any(|record| {
-                record["expr_kind"] == "TypedArrayGet"
-                    && record["consumer"] == "TypedArrayGet.slow_path"
-                    && record["access_mode"] == "dynamic_fallback"
-                    && record["materialization_reason"] == "stale_view_length"
-                    && record["fallback_reason"] == "stale_view_length"
-            }),
-        "expected stale native-owned view length fallback reason:\n{stale_length:#}"
-    );
-
-    let mutable_alias = compile_artifact_json(
-        "artifact_native_owned_mutable_alias.ts",
-        vec![
-            native_arena_owner_let(1, "owner", int(64), false),
-            native_arena_view_let(
-                2,
-                "view",
-                1,
-                "Float64Array",
-                perry_hir::TYPED_ARRAY_KIND_FLOAT64,
-                int(0),
-                int(8),
-            ),
-            Stmt::Let {
-                id: 3,
-                name: "alias".to_string(),
-                ty: Type::Named("Float64Array".to_string()),
-                mutable: false,
-                init: Some(local(2)),
-            },
-            Stmt::Return(Some(index_get(3, int(0)))),
-        ],
-    );
-    assert!(
-        mutable_alias["records"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .any(|record| {
-                record["expr_kind"] == "TypedArrayGet"
-                    && record["consumer"] == "TypedArrayGet.slow_path"
-                    && record["access_mode"] == "dynamic_fallback"
-                    && record["materialization_reason"] == "mutable_alias"
-                    && record["fallback_reason"] == "mutable_alias"
-            }),
-        "expected native-owned mutable alias fallback reason:\n{mutable_alias:#}"
-    );
-
-    let missing_owner = compile_artifact_json(
-        "artifact_native_owned_missing_owner_root.ts",
-        vec![
-            native_arena_owner_let(1, "owner", int(64), true),
-            native_arena_view_let(
-                2,
-                "view",
-                1,
-                "Float64Array",
-                perry_hir::TYPED_ARRAY_KIND_FLOAT64,
-                int(0),
-                int(8),
-            ),
-            Stmt::Expr(Expr::LocalSet(
-                1,
-                Box::new(Expr::NativeArenaAlloc(Box::new(int(64)))),
-            )),
-            Stmt::Return(Some(index_get(2, int(0)))),
-        ],
-    );
-    assert!(
-        missing_owner["records"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .any(|record| {
-                record["expr_kind"] == "TypedArrayGet"
-                    && record["consumer"] == "TypedArrayGet.slow_path"
-                    && record["access_mode"] == "dynamic_fallback"
-                    && record["materialization_reason"] == "missing_owner_root"
-                    && record["fallback_reason"] == "missing_owner_root"
-            }),
-        "expected missing owner-root fallback reason:\n{missing_owner:#}"
-    );
-
-    let escaping = compile_artifact_json(
-        "artifact_native_owned_escaping_pointer.ts",
-        vec![
-            native_arena_owner_let(1, "owner", int(64), false),
-            native_arena_view_let(
-                2,
-                "view",
-                1,
-                "Float64Array",
-                perry_hir::TYPED_ARRAY_KIND_FLOAT64,
-                int(0),
-                int(8),
-            ),
-            Stmt::Expr(extern_call(
-                "escape_native_view",
-                vec![local(2)],
-                Type::Number,
-            )),
-            Stmt::Return(Some(index_get(2, int(0)))),
-        ],
-    );
-    assert!(
-        escaping["records"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .any(|record| {
-                record["expr_kind"] == "TypedArrayGet"
-                    && record["consumer"] == "TypedArrayGet.slow_path"
-                    && record["access_mode"] == "dynamic_fallback"
-                    && record["materialization_reason"] == "escaping_unowned_pointer"
-                    && record["fallback_reason"] == "escaping_unowned_pointer"
-            }),
-        "expected escaping unowned pointer fallback reason:\n{escaping:#}"
-    );
-}
-
-#[test]
-fn uint8_clamped_typed_array_store_records_runtime_fallback() {
-    let body = vec![
-        typed_array_let(
-            1,
-            "clamped",
-            "Uint8ClampedArray",
-            perry_hir::TYPED_ARRAY_KIND_UINT8_CLAMPED,
-            int(8),
-        ),
-        Stmt::Expr(Expr::IndexSet {
-            object: Box::new(local(1)),
-            index: Box::new(int(0)),
-            value: Box::new(number(300.5)),
-        }),
-        Stmt::Return(Some(int(0))),
-    ];
-
-    let artifact = compile_artifact_json("artifact_uint8_clamped_store_fallback.ts", body);
-    assert!(
-        artifact["records"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .any(|record| {
-                record["expr_kind"] == "TypedArraySet"
-                    && record["consumer"] == "TypedArraySet.slow_path"
-                    && record["access_mode"] == "dynamic_fallback"
-                    && !record["fallback_reason"].is_null()
-            }),
-        "expected Uint8ClampedArray store to stay on runtime fallback:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn typed_array_alias_read_records_runtime_fallback() {
-    let body = vec![
-        typed_array_let(
-            1,
-            "array",
-            "Uint16Array",
-            perry_hir::TYPED_ARRAY_KIND_UINT16,
-            int(8),
-        ),
-        Stmt::Let {
-            id: 2,
-            name: "alias".to_string(),
-            ty: Type::Named("Uint16Array".to_string()),
-            mutable: false,
-            init: Some(local(1)),
-        },
-        for_loop(3, int(8), vec![Stmt::Expr(index_get(2, local(3)))]),
-        Stmt::Return(Some(int(0))),
-    ];
-
-    let artifact = compile_artifact_json("artifact_typed_array_alias_fallback.ts", body);
-    assert!(
-        artifact["records"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .any(|record| {
-                record["expr_kind"] == "TypedArrayGet"
-                    && record["consumer"] == "TypedArrayGet.slow_path"
-                    && record["access_mode"] == "dynamic_fallback"
-                    && !record["fallback_reason"].is_null()
-            }),
-        "expected aliased typed-array read to record runtime fallback:\n{artifact:#}"
-    );
-}
-
-#[test]
-fn reassigned_typed_array_store_records_runtime_fallback() {
-    let body = vec![
-        Stmt::Let {
-            id: 1,
-            name: "array".to_string(),
-            ty: Type::Named("Int32Array".to_string()),
-            mutable: true,
-            init: Some(Expr::TypedArrayNew {
-                kind: perry_hir::TYPED_ARRAY_KIND_INT32,
-                arg: Some(Box::new(int(8))),
-            }),
-        },
-        Stmt::Expr(Expr::LocalSet(
-            1,
-            Box::new(Expr::TypedArrayNew {
-                kind: perry_hir::TYPED_ARRAY_KIND_INT32,
-                arg: Some(Box::new(int(8))),
-            }),
-        )),
-        array_set(1, int(0), int(42)),
-        Stmt::Return(Some(index_get(1, int(0)))),
-    ];
-
-    let artifact = compile_artifact_json("artifact_typed_array_reassign_fallback.ts", body);
-    let records = artifact["records"].as_array().unwrap();
-    assert!(
-        records.iter().any(|record| {
-            record["expr_kind"] == "TypedArraySet"
-                && record["consumer"] == "TypedArraySet.slow_path"
-                && record["access_mode"] == "dynamic_fallback"
-                && !record["fallback_reason"].is_null()
-        }),
-        "expected reassigned typed-array store to record runtime fallback:\n{artifact:#}"
-    );
-    assert!(
-        records.iter().any(|record| {
-            record["expr_kind"] == "TypedArrayGet"
-                && record["consumer"] == "TypedArrayGet.slow_path"
-                && record["access_mode"] == "dynamic_fallback"
-                && !record["fallback_reason"].is_null()
-        }),
-        "expected reassigned typed-array read to record runtime fallback:\n{artifact:#}"
     );
 }
